+++ packages/zent-form/src/form_components/InputField.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import Input from 'zent-input';
 import getControlGroup from '../getControlGroup';
+import unkownProps from '../unkownProps';
+import omit from 'zent-utils/lodash/omit';
 
-const InputField = getControlGroup(({ type = 'text', ...passableProps }) => {
+const InputField = getControlGroup(({ type = 'text', ...rest }) => {
+  const passableProps = omit(rest, unkownProps);
   return <Input {...passableProps} type={type} />;
 });