+++ packages/zent-form/src/form_components/RadioGroupField.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import Radio from 'zent-radio';
 import getControlGroup from '../getControlGroup';
+import unkownProps from '../unkownProps';
+import omit from 'zent-utils/lodash/omit';
 
 const RadioGroup = Radio.Group;
 const RadioGroupField = getControlGroup(props => {
-  return <RadioGroup className="zent-form__radio-group" {...props} />;
+  const passableProps = omit(props, unkownProps);
+  return <RadioGroup className="zent-form__radio-group" {...passableProps} />;
 });
 
 export default RadioGroupField;